[#213] Add semantic string tests

+++ iceoryx2-bb/container/src/byte_string.rs
@@ -49,7 +49,7 @@ pub unsafe fn strnlen(ptr: *const core::ffi::c_char, len: usize) -> usize {
     const NULL_TERMINATION: core::ffi::c_char = 0;
     for i in 0..len {
         if *ptr.add(i) == NULL_TERMINATION {
-            return i as usize;
+            return i;
         }
     }
